Sprint and dash ability created through abilities tied to input.
Dash queries area in a line in front of player (or other given direction, find the furthest valid position, end will motion warp the dash animations root motion to move the player to that position in time with the animation. Also passes throuhg objects with collisions that do not block the given collision type.

Sprint applied the gameplay tag to the player while the input is held. The tag indicates to the player controller to run a different block of code for movement input and speed to the character movement controller.

TODO: Modify character to be programming driven rather than root motion through animation. Better to lean on responsiveness rather than 100% smooth animation